# Show traffic split metrics from sources in all namespaces